Refactor commands
* Split up the core commands

* Make geoip a script
