+++ piqueserver/commands/__init__.py
@@ -1,0 +1,2 @@
+from .command import * # for script compatibility
+from .core_commands import * # to register the core commands
